refactor(#1): Issue List Item 컴포넌트 분리
