+++ src/components/IssueList.tsx
@@ -1,15 +1,9 @@
-import { Fragment, ReactNode } from 'react';
-import { Link } from 'react-router-dom';
-import { GitHubIssue } from 'github';
-
+import { Fragment } from 'react';
 import { useGitHubIssueList } from 'contexts/gitHubIssueListContext';
 import useIntersect from 'hooks/useIntersect';
 import colors from 'constants/colors';
-import routes from 'constants/routes';
-import convertToKoreanDate from 'utils/convertToKoreanDate';
-
-import { Skeleton } from './Skeleton';
 import { WantedAdvertisementImage } from './WantedAdvertisementImage';
+import { IssueListItem, IssueListItemSkeleton } from './IssueListItem';
 
 export function IssueList() {
   const { issueList, fetchNextIssueList, isLoading } = useGitHubIssueList();
@@ -25,11 +19,11 @@ export function IssueList() {
                 display: 'block',
                 textAlign: 'center',
                 padding: '1rem 0',
-                borderBottom: `1px solid ${BORDER_COLOR}`,
+                borderBottom: `1px solid ${colors.border}`,
               }}
             />
           )}
-          <IssueItem {...gitHubIssue} />
+          <IssueListItem hasLink {...gitHubIssue} />
         </Fragment>
       ))}
       {isLoading ? <IssueListSkeleton /> : <div ref={intersectRef} css={{ height: '1px' }} />}
@@ -38,74 +32,5 @@ export function IssueList() {
 }
 
 function IssueListSkeleton() {
-  return Array.from({ length: 30 }, (_, i) => (
-    <IssueItemTemplate
-      key={i}
-      title={<Skeleton css={{ width: '100%', height: '2rem' }} />}
-      detail={<Skeleton css={{ width: '50%', height: '1.5rem' }} />}
-      comment={<Skeleton css={{ width: '4.5rem', height: '1.5rem' }} />}
-    />
-  ));
+  return Array.from({ length: 30 }, (_, i) => <IssueListItemSkeleton key={i} />);
 }
-
-function IssueItem({ number, title, comments, created_at, user }: GitHubIssue) {
-  return (
-    <IssueItemTemplate
-      title={
-        <Link
-          to={`${routes.issues}/${number}`}
-          css={{
-            cursor: 'pointer',
-            fontSize: FONT_SIZE.large,
-            fontWeight: 600,
-            ':hover': { color: colors.blue400 },
-          }}
-        >
-          <span css={{ color: colors.grey600 }}>#{number}</span> {title}
-        </Link>
-      }
-      detail={
-        <div css={{ fontSize: FONT_SIZE.medium }}>
-          <h3 css={{ display: 'inline-block', marginRight: '0.3rem' }}>작성자: {user.login},</h3>
-          <time dateTime={created_at}>작성일: {convertToKoreanDate(created_at)}</time>
-        </div>
-      }
-      comment={<>코멘트: {comments}</>}
-    />
-  );
-}
-
-interface IssueItemTemplateProps {
-  title: ReactNode;
-  detail: ReactNode;
-  comment: ReactNode;
-}
-
-function IssueItemTemplate({ title, detail, comment }: IssueItemTemplateProps) {
-  return (
-    <li
-      css={{
-        lineHeight: 1.3,
-        padding: '1rem 0',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        gap: '7%',
-        borderBottom: `1px solid ${BORDER_COLOR}`,
-      }}
-    >
-      <div css={{ display: 'flex', flexDirection: 'column', gap: '0.7rem', flexGrow: 1 }}>
-        {title}
-        {detail}
-      </div>
-      <div css={{ fontSize: FONT_SIZE.medium, textAlign: 'end', whiteSpace: 'nowrap' }}>{comment}</div>
-    </li>
-  );
-}
-
-const BORDER_COLOR = colors.grey200;
-
-const FONT_SIZE = {
-  medium: '0.9rem',
-  large: '1.1rem',
-};
