+++ src/pages/IssueListPage.tsx
@@ -1,11 +1,15 @@
 import { Header } from 'components/Header';
 import { IssueList } from 'components/IssueList';
 import { GitHubIssueListProvider } from 'contexts/gitHubIssueListContext';
-
-const owner = 'facebook';
-const repo = 'react';
+import { useParams } from 'react-router-dom';
 
 export function IssueListPage() {
+  const { owner, repo } = useParams<{ owner: string; repo: string }>();
+
+  if (!owner || !repo) {
+    return null;
+  }
+
   return (
     <>
       <Header owner={owner} repo={repo} />
